Split acceleration code into init and runtime to save flash space
